refactor: コンポーネントのリネーム
